extending ERC20 tokens

+++ contracts/Token.sol
@@ -3,23 +3,25 @@ pragma solidity ^0.8.4;
 
 import "hardhat/console.sol";
 
+// import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
+
 contract Token {
-  string public name = "Nader Dabit Token";
-  string public symbol = "NDT";
-  uint public totalSupply = 1000000;
-  mapping(address => uint) balances;
+    string public name = "Nader Dabit Token";
+    string public symbol = "NDT";
+    uint256 public totalSupply = 1000000;
+    mapping(address => uint256) balances;
 
-  constructor() {
-    balances[msg.sender] = totalSupply;
-  }
+    constructor() {
+        balances[msg.sender] = totalSupply;
+    }
 
-  function transfer(address to, uint amount) external {
-    require(balances[msg.sender] >= amount, "Not enough tokens");
-    balances[msg.sender] -= amount;
-    balances[to] += amount;
-  }
+    function transfer(address to, uint256 amount) external {
+        require(balances[msg.sender] >= amount, "Not enough tokens");
+        balances[msg.sender] -= amount;
+        balances[to] += amount;
+    }
 
-  function balanceOf(address account) external view returns (uint) {
-    return balances[account];
-  }
+    function balanceOf(address account) external view returns (uint256) {
+        return balances[account];
+    }
 }
